added _trueNode and _falseNode to resolve issue 5
This fix resolves the error noted in issue 5 by adding additional
variables to hold the trueNode and falseNode dom elements. This doesn't
get to the root of the issue, i.e., why do we loose a reference to
this.trueNode and this.falseNode during a callback from a registered
attribute listener?